chore: move helpers to utils
